Pointer Variable
A pointer variable is a variable which gives the memory address value of a stored variable. It is always an integer type data. ' * ' is used to get the value of that memory location. Two pointers can be subtracted but other arithmetic operations are not allowed. size of any pointer is 2 byte.